+++ src/Language/GraphQL/Validate/Validation.hs
@@ -52,6 +52,7 @@ data Rule m
     | SelectionRule (Selection -> RuleT m)
     | FragmentRule (FragmentDefinition -> RuleT m) (InlineFragment -> RuleT m)
     | FragmentSpreadRule (FragmentSpread -> RuleT m)
+    | FieldRule (Field -> RuleT m)
 
 -- | Monad transformer used by the rules.
 type RuleT m = ReaderT (Validation m) Maybe Error